Core function to go from a vector of species, newick string, or phylo object get a chronogram or dates back

Core function to go from a vector of species, newick string, or phylo object get a chronogram or dates back

EstimateDates(input = c("Rhea americana", "Pterocnemia pennata",
  "Struthio camelus"), output.format = "phylo.sdm", partial = TRUE,
  usetnrs = FALSE, approximatematch = TRUE,
  cache = get("opentree_chronograms"), method = "PATHd8")

Arguments

input
A vector of names, a newick string, or a phylo object

output.format
The desired output format. See details.

partial
If TRUE, use source trees even if they only match some of the desired taxa

usetnrs
If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer

approximatematch
If TRUE, use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches)

cache
The cached set of chronograms and other info from data(opentree_chronograms)

method
The method used for congruification. PATHd8 only right now, r8s and treePL later.

Value

Varies depending on the chosen format

Details

The output formats are citations, mrca, newick.all, newick.sdm, newick.median, phylo.sdm, phylo.median, phylo.all, html

Examples

ages <- EstimateDates(c("Rhea americana", "Pterocnemia pennata", "Struthio camelus", "Mus musculus"), output.format="mrca")

#> [1] "Using trees from:"
#> [1] "Grealy, Alicia, Matthew Phillips, Gifford Miller, M. Thomas P. Gilbert, Jean-Marie Rouillard, David Lambert, Michael Bunce, James Haile. 2017. Eggshell palaeogenomics: Palaeognath evolutionary history revealed through ancient nuclear and mitochondrial DNA from Madagascan elephant bird (Aepyornis sp.) eggshell. Molecular Phylogenetics and Evolution 109: 151-163"
#> [2] "Hedges, S. Blair, Julie Marin, Michael Suleski, Madeline Paymer, Sudhir Kumar. 2015. Tree of life reveals clock-like speciation and diversification. Molecular Biology and Evolution 32 (4): 835-845"                                                                                                                                                                      
#> [3] "Hedges, S. Blair, Julie Marin, Michael Suleski, Madeline Paymer, Sudhir Kumar. 2015. Tree of life reveals clock-like speciation and diversification. Molecular Biology and Evolution 32 (4): 835-845"                                                                                                                                                                      
#> [4] "Prum, Richard O., Jacob S. Berv, Alex Dornburg, Daniel J. Field, Jeffrey P. Townsend, Emily Moriarty Lemmon, Alan R. Lemmon. 2015. A comprehensive phylogeny of birds (Aves) using targeted next-generation DNA sequencing. Nature 526, (7574): 569-573"                                                                                                                   
#> [5] "Prum, Richard O., Jacob S. Berv, Alex Dornburg, Daniel J. Field, Jeffrey P. Townsend, Emily Moriarty Lemmon, Alan R. Lemmon. 2015. A comprehensive phylogeny of birds (Aves) using targeted next-generation DNA sequencing. Nature 526, (7574): 569-573"                                                                                                                   
#> [6] "Claramunt, Santiago, Joel Cracraft. 2015. A new time tree reveals Earth historys imprint on the evolution of modern birds. Science Advances 1 (11): e1501005-e1501005"                                                                                                                                                                                                     
#> [7] "Jetz, W., G. H. Thomas, J. B. Joy, K. Hartmann, A. O. Mooers. 2012. The global diversity of birds in space and time. Nature 491 (7424): 444-448"                                                                                                                                                                                                                           
#> [8] "Jetz, W., G. H. Thomas, J. B. Joy, K. Hartmann, A. O. Mooers. 2012. The global diversity of birds in space and time. Nature 491 (7424): 444-448"                                                                                                                                                                                                                           
#> [9] "Mitchell, K. J., B. Llamas, J. Soubrier, N. J. Rawlence, T. H. Worthy, J. Wood, M. S. Y. Lee, A. Cooper. 2014. Ancient DNA reveals elephant birds and kiwi are sister taxa and clarifies ratite bird evolution. Science 344 (6186): 898-900."