Function from Christoph Heibl's phyloch package Add a Geological Time Axis for Time-Calibrated Phylogenies.

Function from Christoph Heibl's phyloch package Add a Geological Time Axis for Time-Calibrated Phylogenies.

axisGeo(
  GTS,
  tip.time = 0,
  unit = c("epoch", "period"),
  ages = TRUE,
  cex = 1,
  col = "white",
  texcol = "black",
  gridty = 0,
  gridcol = "black"
)

Arguments

GTS
a data frame containing the geological time scale such as provided by strat2012.

tip.time
a real number giving time units back from present were the youngest taxa were sampled.

unit
a vector of mode character used to select geological time units that shall be displayed. When using gradstein04, "eon", "era", "period", "epoch", and "stage" are available.

ages
logical: if TRUE, a real-numbered axis of unit boundaries is added to the plot.

cex
a real number giving the character expansion.

col
a vector of mode character used to select background color(s). Will be recycled if necessary.

texcol
a vector of mode character used to select font color(s). Will be recycled if necessary.

gridty
an integer representing the line type of the grid exactly in the same way as lty argument in par: 0 or blank, 1 or "solid", 2 or "dashed", 3 or "dotted", 4 or "dotdash", 5 or "longdash", and 6 or "twodash".

gridcol
a vector of mode character used to select grid color(s). Will be recycled if necessary.

Value

Adds geological time scale to a phylo plot.

Details

Make sure ape is loaded otherwise it won't find .PlotPhyloEnv If the name of a time interval does not fit the space on the plotting device provided by that time frame, axisGeo tries to fit the name string by lowering the original character expansion. In order to avoid undiscernable small font sizes only character expansion of 50 percent or more of the original cex is allowed; in all other cases strings are abbreviated and plotted with the original character expansion.

References

Gradstein F.M., J.G. Ogg & A.G. Smith. 2004. A Geologic Time Scale. Cambridge University Press, Cambridge, UK. www.cambridge.org/uk/catalogue/catalogue.asp?isbn=0521786738. International Commission on Stratigraphy. 2012. International stratigraphic chart: http://www.stratigraphy.org/upload/ISChart2009.pdf.

Author

Christoph Heibl