Choose an ultrametric phylo object from cluster_patristicmatrix() obtained with a particular clustering method, or the next best tree. If there are no ultrametric trees, it does not force them to be ultrametric.

Choose an ultrametric phylo object from cluster_patristicmatrix() obtained with a particular clustering method, or the next best tree. If there are no ultrametric trees, it does not force them to be ultrametric.

Usage

choose_cluster(phycluster, clustering_method = "nj")

Arguments

phycluster

An output from cluster_patristicmatrix()

clustering_method

A character vector indicating the method to construct the tree. Options are:

nj

Neighbor-Joining method applied with ape::nj().

upgma

Unweighted Pair Group Method with Arithmetic Mean method applied with phangorn::upgma().

bionj

An improved version of the Neighbor-Joining method applied with ape::bionj().

triangle

Triangles method applied with ape::triangMtd()

mvr

Minimum Variance Reduction method applied with ape::mvr().

Value

A phylo object or NA.