Identifies, extracts and cleans children names from a taxonomy_taxon_info() output

Identifies, extracts and cleans children names from a taxonomy_taxon_info() output

clean_taxon_info_children(
  taxon_info,
  invalid = c("barren", "extinct", "uncultured", "major_rank_conflict",
    "incertae_sedis", "unplaced", "conflict", "environmental", "not_otu", "hidden",
    "hybrid")
)

Arguments

taxon_info
An output of rotl::taxonomy_taxon_info()

invalid
A character vector of flags used by Open Tree Taxonomy to detect invalid taxon names

Value

a list with valid children unique OTT names, ott_ids and ranks

Details

eliminates all taxa that will give problems when trying to retrieve an induced subtree from otol.

Examples

# some suppressed taxa within "Gleicheniales" family of ferns:
tt <- rotl::taxonomy_taxon_info(3866048)
rotl::taxonomy_taxon_info(5262766)

#> $`5262766`
#> $`5262766`$flags
#> $`5262766`$flags[[1]]
#> [1] "hybrid"
#> 
#> 
#> $`5262766`$is_suppressed
#> [1] FALSE
#> 
#> $`5262766`$is_suppressed_from_synth
#> [1] TRUE
#> 
#> $`5262766`$name
#> [1] "Sticherus cunninghamii x Sticherus urceolatus"
#> 
#> $`5262766`$ott_id
#> [1] 5262766
#> 
#> $`5262766`$rank
#> [1] "species"
#> 
#> $`5262766`$source
#> [1] "ott3.3draft1"
#> 
#> $`5262766`$synonyms
#> $`5262766`$synonyms[[1]]
#> [1] "Sticherus cunninghamii x S. urceolatus"
#> 
#> 
#> $`5262766`$tax_sources
#> $`5262766`$tax_sources[[1]]
#> [1] "ncbi:1280269"
#> 
#> 
#> $`5262766`$unique_name
#> [1] "Sticherus cunninghamii x Sticherus urceolatus"
#> 
#> 
#> attr(,"class")
#> [1] "taxon_info"