Eliminates unmatched (NAs) and invalid taxa from a rotl::tnrs_match_names or datelife::tnrs_match output Useful to get ott ids to retrieve an otol induced subtree from. Needed because using include_suppressed = FALSE in tnrs_match_names does not drop all invalid taxa

Eliminates unmatched (NAs) and invalid taxa from a rotl::tnrs_match_names or datelife::tnrs_match output Useful to get ott ids to retrieve an otol induced subtree from. Needed because using include_suppressed = FALSE in tnrs_match_names does not drop all invalid taxa

clean_tnrs(
  tnrs,
  invalid = c("barren", "extinct", "uncultured", "major_rank_conflict", "incertae",
    "unplaced", "conflict", "environmental", "not_otu"),
  remove_nonmatches = FALSE
)

Arguments

tnrs
A data frame, usually an output from datelife::tnrs_match or rotl::tnrs_match_names functions, but see details.

invalid
A character string with flags to be removed from final object.

remove_nonmatches
Boolean, whether to remove unsuccessfully matched names or not.

Value

A data frame or named list (depending on the input) with valid taxa only.

Details

Input can be any data frame or named list that relates taxa stored in an element named "unique" to a validity category stored in "flags".