Congruify and Check.

Congruify and Check.

Usage

congruify_and_check(
  reference,
  target,
  taxonomy = NULL,
  tol = 0.01,
  option = 2,
  scale = "pathd8",
  attempt_fix = TRUE
)

Arguments

reference

an ultrametric tree used to time-scale the target

target

a phylogram that is sought to be ultrametricized based on the reference phylogeny

taxonomy

a linkage table between tips of the phylogeny and clades represented in the tree; rownames of 'taxonomy' should be tips found in the phylogeny

tol

branching time in reference above which secondary constraints will be applied to target

option

an integer (1 or 2; see details).

scale

NA, "PATHd8" or "treePL" (if PATHd8 or "treePL" are available in the R PATH)

attempt_fix

Default to TRUE. If congruification results in NA branch lengths, it will attempt to fix them.