Information on contributors, authors, study ids and clades from studies with chronograms in Open Tree of Life (Open Tree)

Information on contributors, authors, study ids and clades from studies with chronograms in Open Tree of Life (Open Tree)

Usage

contributor_cache

Format

A list of five data sets.

author.pretty

A character vector with the author names from studies with chronograms that are in Open Tree.

author.results

A dataframe with three variables: authors, study ids and clades.

curator.pretty

A character vector with the names of curators of chronograms that are in Open Tree.

curator.results

A data.frame with three variables: curators, study ids and clades.

missed_doi

A character vector with study ids whose "doi" could not be retrieved.

Source

http://opentreeoflife.org

Details

Generated with make_contributor_cache().