Date with Paleobiology Database and paleotree.

This will take a topology, look up information about fossils for taxa on the tree, and use paleotree::timePaleoPhy() to compute branch lengths.

Usage

date_with_pbdb(phy, recent = FALSE, assume_recent_if_missing = TRUE)

Arguments

phy

A phylo object.

recent

If TRUE, forces the minimum age to be zero for any taxon

assume_recent_if_missing

If TRUE, any taxon missing from PBDB is assumed to be recent.

Value

A dated tree.

Examples

if (FALSE)  # This is a flag for package development. You are welcome to run the example.

taxa <- c(
  "Archaeopteryx", "Pinus", "Quetzalcoatlus", "Homo sapiens",
  "Tyrannosaurus rex", "Megatheriidae", "Metasequoia", "Aedes", "Panthera"
)
phy <- tree_from_taxonomy(taxa, sources = "The Paleobiology Database")$phy
#> Warning: `gnr_datasources()` was deprecated in taxize v0.9.102.
#> ℹ Please use `gna_data_sources()` instead.
#> ℹ The deprecated feature was likely used in the datelife package.
#>   Please report the issue to the authors.
#> Error: Not Found (HTTP 404)
 # \dontrun{} # end dontrun