Function to compute the SDM supertree (Criscuolo et al. 2006) from a datelifeResult object.

Function to compute the SDM supertree (Criscuolo et al. 2006) from a datelifeResult object.

datelife_result_sdm(datelife_result, weighting = "flat", verbose = TRUE, ...)

Arguments

datelife_result
An output of get_datelife_result function: A list of patristic matrices with names corresponding to original study citations.

weighting
A character vector indicating how much weight to give to each input tree in the SDM analysis. Choose one of:

weighting = "flat"

All trees have equal weighting.

weighting = "taxa"

Weight is proportional to number of taxa.

weighting = "inverse"

Weight is proportional to 1 / number of taxa.

verbose
Boolean. If TRUE, it gives printed updates to the user.

...
Arguments passed on to summary_matrix_to_phylo

summ_matrix

A summary patristic distance matrix from sdm or median. See details.

total_distance

Boolean. If TRUE it will divide the matrix in half, if FALSE it will take it as is.

use

A character vector indicating what type of age to use for summary. One of the following

mean

It will use the mean of the node age distributions.

min

It will use the minimum age from the node age distributions.

max

Choose this if you wanna be conservative; it will use the maximum age from the node age distributions.

target_tree

A phylo object. Use this in case you want a particular backbone for the output tree.

datelife_query

A datelifeQuery object, output of make_datelife_query function

Value

A phylo object from SDM with an extra element:

data

A datelifeResult object with the chronograms that were used to construct the SDM tree.

Details

Criscuolo A, Berry V, Douzery EJ, Gascuel O. SDM: a fast distance-based approach for (super) tree building in phylogenomics. Syst Biol. 2006. 55(5):740. doi: 10.1080/10635150600969872.