Extract secondary calibrations from a given datelifeResult object

This function extracts node ages for each taxon pair given as input. It applies the congruification method described in Eastman et al. (2013) doi: 10.1111/2041-210X.12051 , to create a matrix of suitable secondary calibrations for each taxon pair. Congruification is implemented with the function geiger::congruify.phylo().

Usage

extract_calibrations_dateliferesult(input = NULL, each = FALSE)

Arguments

input

A datelifeResult object.

each

Boolean, default to FALSE: all calibrations are returned in the same data.frame. If TRUE, calibrations from each chronogram are returned in separate data frames.

Value

An object of class datelifeCalibrations, i.e., a data.frame (if each = FALSE) or a list of data.frames (if each = TRUE) of secondary calibrations, for each pair of taxon names in input. The attribute chronograms stores the input data from which the calibrations were extracted.

Details

The function calls summarize_datelife_result()] with summary_format = "phylo_all" to go from a datelifeResultobject to aphyloormultiPhylo` object that will be passed to extract_calibrations_phylo().