Use congruification to extract secondary calibrations from a phylo or multiPhylo object with branch lengths proportional to time.

This function extracts node ages for each taxon pair given in input$tip.labels. It applies the congruification method described in Eastman et al. (2013) doi:10.1111/2041-210X.12051 , implemented with the function geiger::congruify.phylo(), to create a data.frame of taxon pair node ages that can be used as secondary calibrations.

Usage

extract_calibrations_phylo(input = NULL, each = FALSE)

Arguments

input

A phylo or multiPhylo object with branch lengths proportional to time.

each

Boolean, default to FALSE: all calibrations are returned in the same data.frame. If TRUE, calibrations from each chronogram are returned in separate data frames.

Value

An object of class calibrations, which is a data.frame (if each = FALSE) or a list of data.frames (if each = TRUE) of node ages for each pair of taxon names. You can access the input data from which the calibrations were extracted with attributes(output)$chronograms.

References

Eastman et al. (2013) "Congruification: support for time scaling large phylogenetic trees". Methods in Ecology and Evolution, 4(7), 688-691, doi:10.1111/2041-210X.12051 .