Get all calibrations available for a set of taxa.

Get all calibrations available for a set of taxa.

get_all_calibrations(
  input = c("Rhea americana", "Pterocnemia pennata", "Struthio camelus"),
  partial = TRUE,
  use_tnrs = FALSE,
  approximate_match = TRUE,
  update_cache = FALSE,
  cache = getAnywhere("opentree_chronograms"),
  verbose = FALSE,
  each = FALSE
)

Arguments

input
vector of names, a newick string, a phylo or multiPhylo object with branch lengths proportional to time, a datelifeResult object

partial
Boolean; default TRUE: use source trees even if they only match some of the desired taxa

use_tnrs
Boolean; default False. If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer

approximate_match
Boolean; default TRUE: use a slower TNRS to correct misspellings, increasing the chance of matches (including false matches)

update_cache
default to FALSE

cache
The cached set of chronograms and other info from data(opentree_chronograms)

verbose
Boolean. If TRUE, it gives printed updates to the user.

each
Boolean, default to FALSE all calibrations are returned in the same data frame. If TRUE, calibrations from each chronogram are returned in a separate data frame.

Value

A data frame (or list of data frames if each = TRUE) of calibrations, with attribute "chronograms" containing the dated trees that were used to get calibrations from.