Get a dated OpenTree induced synthetic subtree from a set of given taxon names, from blackrim's FePhyFoFum service.

Get a dated OpenTree induced synthetic subtree from a set of given taxon names, from blackrim's FePhyFoFum service.

Usage

get_dated_otol_induced_subtree(input = NULL, ott_ids = NULL, ...)

Arguments

input

Optional. A character vector of names or a datelifeQuery object.

ott_ids

If not NULL, it takes this argument and ignores input. A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info() or rotl::tnrs_match_names() or tnrs_match().

...

Arguments passed on to check_ott_input

Value

A phylo object with edge length proportional to time in Myrs. It will return NA if any ott_id is invalid.

Details

OpenTree dated tree from Stephen Smith's OpenTree scaling service at https://github.com/FePhyFoFum/gophy if you want to make an LTT plot of a dated OpenTree tree you'll need to get rid of singleton nodes with ape::collapse.singles() and also probably do phytools::force.ultrametric().