Get all chronograms from Open Tree of Life database using direct call from Open Tree API

Get all chronograms from Open Tree of Life database using direct call from Open Tree API

Usage

get_opentree_chronograms(max_tree_count = "all")

get_otol_chronograms(max_tree_count = "all")

Arguments

max_tree_count

Default to "all", it gets all available chronograms. For testing purposes, a numeric value indicating the max number of trees to be cached.

Value

A list of 4 elements:

authors

A list of lists of author names of the original studies that published chronograms currently stored in the Open Tree of Life database.

curators

A list of lists of curator names that uploaded chronograms to the Open Tree of Life database.

studies

A list of study identifiers from original studies that published chronograms currently stored in the Open Tree of Life database.

trees

A multiPhylo object storing the chronograms from Open Tree of Life database.

update

A character vector indicating the time when the database object was last updated.

version

A character vector indicating the datelife package version when the object was last updated.