Takes a constraint tree and uses mrBayes to get node ages and branch lengths given a set of node calibrations without any data.

Takes a constraint tree and uses mrBayes to get node ages and branch lengths given a set of node calibrations without any data.

make_mrbayes_tree(
  constraint = NULL,
  taxa = NULL,
  ncalibration = NULL,
  missing_taxa = NULL,
  age_distribution = "fixed",
  root_calibration = FALSE,
  mrbayes_output_file = "mrbayes_run.nexus"
)

Arguments

constraint
The constraint tree: a phylo object or a newick character string, with or without branch lengths.

taxa
A character vector with taxon names to be maintained in tree

ncalibration
The node calibrations: a phylo object with branch lengths proportional to time; in this case all nodes from ncalibration will be used as calibration points. Alternatively, a list with two elements: the first is a character vector with node names from phy to calibrate; the second is a numeric vector with the corresponding ages to use as calibrations.

missing_taxa
A tree, a data frame or a vector enlisting all missing taxa you want to include.

A tree.

Either as a phylo object or as a newick character string. It contains all taxa that you want at the end, both missing and non missing. This tree will be used as a hard constraint.

A data frame.

It contains two columns. taxon. A character vector of missing taxon names. clade. A character or numeric vector of nodes from a constraint tree to which each taxon will be assigned.

A character vector.

It contains the names of the missing taxa. They will be added at random to the constraint tree.

age_distribution
A character string specifying the type of calibration. Only "fixed" and "uniform" are implemented for now.

fixed

The age given in ncalibration will be used as fixed age.

lognormal

The age given in ncalibration will be used as mean age. The standard deviation can be provided. # still need to add this option. By default, a 95 CI sd is used.

uniform

The age given in ncalibration will be used as mean age. Where min_age = 0.9 * mean age, and max_age = 1.1 * mean age.

root_calibration
Used to set a calibration at the root or not. Default to FALSE. Only relevant if ncalibration is specified.

mrbayes_output_file
A character vector specifying the name of mrBayes run file and outputs (can specify directory too).

Value

A phylo tree with branch lengths proportional to time. It will save all mrBayes outputs in the working directory.