Chronogram database

Now storing >200 chronograms from Open Tree of Life

opentree_chronograms

Format

A list of four elements, containing data from OpenTree of Life chronograms

authors

List of lists of authors of the included studies

curators

List of lists of curators of the included studies

studies

List of study identifiers

trees

List storing the chronograms from OpenTree

Source

http://opentreeoflife.org

Details

Generated with opentree_chronograms <- get_otol_chronograms() usethis::use_data(opentree_chronograms, overwrite = T) and updated with update_datelife_cache()