Easy visualization of lineage through time plots of all chronograms in a multiphylo object.

Easy visualization of lineage through time plots of all chronograms in a multiphylo object.

plot_ltt_phyloall(
  taxon = NULL,
  phy,
  ltt_colors = NULL,
  tax_datedotol = NULL,
  file_name = NULL,
  file_dir = NULL,
  height = 3.5,
  width = 7,
  add_legend = FALSE,
  add_title = FALSE,
  title_text = NULL,
  study_number_cex = 0.75,
  lwd_arrows = 2,
  ...
)

Arguments

taxon
Character vector indicating the name of the taxon or lineage that the chronograms in phy belong to.

phy
A phylo or multiphylo object with chronograms (trees with branch lengths proportional to geologic time), ideally.

ltt_colors
A character vector indicating the colors to be used for plotting ltt

tax_datedotol
A chronogram to compare trees in phy to.

file_name
A character string giving the name of the pdf file.

file_dir
A character string giving the path to write the file to.

height
Height of the plot output

width
width of the plot output

add_legend
Boolean

add_title
Boolean

title_text
Character vector

study_number_cex
A numeric value

lwd_arrows
A numeric value

...
Arguments passed on to ape::ltt.plot

xlab

a character string (or a variable of mode character) giving the label for the \(x\)-axis (default is "Time").

ylab

idem for the \(y\)-axis (default is "N").

backward

a logical value: should the time axis be traced from the present (the default), or from the root of the tree?

tol

a numeric value (see details).