Core function to generate results

Core function to generate results

Usage

run(
  input = c("Rhea americana", "Pterocnemia pennata", "Struthio camelus"),
  format = "citations",
  partial = "yes",
  plot.width = 600,
  plot.height = 600,
  use_tnrs = "no",
  opentree_chronograms = NULL
)

Arguments

input

A newick string or vector of taxa

format

The output format

partial

How to deal with trees that have a subset of taxa in the query

plot.width

Width in pixels for output plot

plot.height

Height in pixels for output plot

use_tnrs

Whether to use OpenTree's TNRS for the input

opentree_chronograms

The list of lists containing the input trees and other info

Value

results in the desired format