Get summary statistics of ages in a congruifiedCalibrations object.

Function summarize_congruifiedCalibrations returns a table of summary statistics for each node in congruified_calibrations argument.

Usage

summarize_congruifiedCalibrations(congruified_calibrations, age_column)

Arguments

congruified_calibrations

A congruifiedCalibrations object, output of congruify_and_mrca_multiPhylo().

age_column

A character string indicating the name of the column to be summarized.

Value

A data.frame of summarized ages.