Summarize a matchedCalibrations object summary.matchedCalibrations gets the node age distribution from a matchedCalibrations object.

Summarize a matchedCalibrations object summary.matchedCalibrations gets the node age distribution from a matchedCalibrations object.

Usage

# S3 method for class 'matchedCalibrations'
summary(object, ...)

Arguments

object

A matchedCalibrations object, usually an element of the output of match_all_calibrations().

...

Further arguments passed to or from other methods.

Value

A summaryMatchedCalibrations object, which is a list of two matchedCalibrations objects:

not_in_phy

A data.frame subset of input matchedCalibrations object containing taxon name pairs that were not present in the given tree. NULL if all input taxon names are found in the given tree.

in_phy

A data.frame subset of input matchedCalibrations object containing all taxon name pairs that were present in the given tree.

Details

Columns in_phy$mrca_node_name and in_phy$reference are factors.