Date a given tree topology using a combined set of given calibrations

use_calibrations combines all given calibrations and uses them as constraints to perform a dating analysis on a given tree topology, using BLADJ if it has no branch lengths, or PATHd8 if the given tree topology has initial branch lengths.

use_calibrations(
  phy = NULL,
  calibrations = NULL,
  dating_method = "bladj",
  type = "median",
  ...
)

Arguments

phy

A phylo object to use as tree topology.

calibrations

A datelifeCalibrations object, an output of get_all_calibrations().

dating_method

Tree dating method to use. Options are "bladj" or "pathd8".

type

The type of age to use as calibration: "median", "mean", "min", or "max".

...

Arguments passed on to use_calibrations_pathd8

expand

How much to expand by each step to get consistent calibrations. Should be between 0 and 1.

giveup

How many expansions to try before giving up

Value

A phylo object with branch lengths proportional to time.

Details

calibrations and dating_method are stored as attributes. They can be accessed with attributes(return)$datelife_calibrations and attributes(return)$dating_method, respectively.
If input tree does not have branch lengths, dating_method is ignored and the function will use secondary calibrations to date the input tree with BLADJ from phylocom, using phylocomr::ph_bladj(), via use_calibrations_bladj(). If input tree has branch lengths it can be dated with PATHd8, using geiger::PATHd8.phylo(), via use_calibrations_pathd8().