Use calibrations to date a topology with the BLADJ algorithm.

The function prepares the input for BLADJ and calls make_bladj_tree()

Usage

use_calibrations_bladj.matchedCalibrations(
  calibrations,
  type = "mean",
  root_age = NULL
)

Arguments

calibrations

A data.frame of secondary calibrations for any pair of taxon names in phy, usually obtained with get_all_calibrations().

type

The type of age to use as calibration. Options are "median", "mean", "min", or "max".

root_age

Numeric specifying an age for the root, provided by the user. Only used if there are no time calibrations for the root node in the chronograms database. If NULL or not numeric, the maximum calibration age plus one unit of the sd (calculated with stats::sd()) of all node ages available for the tree will be used as root calibration. If there is only one calibration available for the whole tree, the root node age will be proportional to 1.1 of the age of that calibration.

Value

A phylo object with branch lengths proportional to time.

Details

The BLADJ algorithm is part of the Phylocom software, presented in Webb et al. (2008) doi:10.1093/bioinformatics/btn358 .

References

Webb, C. O., Ackerly, D. D., & Kembel, S. W. (2008). "Phylocom: software for the analysis of phylogenetic community structure and trait evolution". Bioinformatics, 24(18), doi:10.1093/bioinformatics/btn358 .