Date a given tree topology by using a given list of calibrations independently, to generate multiple hypothesis of time of divergence

use_calibrations_each wraps use_calibrations to take each set of given calibrations and use it independently as constraints for BLADJ or PATHd8 to date a given tree topology.

Usage

use_calibrations_each(phy = NULL, calibrations = NULL, ...)

Arguments

phy

A phylo object to use as tree topology.

calibrations

A calibrations object, an output of get_all_calibrations().

...

Arguments passed on to use_calibrations

dating_method

Tree dating algorithm to use. Options are "bladj" or "pathd8" (Webb et al., 2008, doi:10.1093/bioinformatics/btn358 ; Britton et al., 2007, doi:10.1080/10635150701613783 ).

type

The type of age to use as calibration. Options are "median", "mean", "min", or "max".

Value

A multiPhylo object of trees with branch lengths proportional to time.

Details

If phy has no branch lengths, dating_method is ignores, and the function applies secondary calibrations to date the tree with the BLADJ algorithm. See make_bladj_tree() and use_calibrations_bladj(). If phy has branch lengths, the function can use the PATHd8 algorithm. See use_calibrations_pathd8().

More

The output object stores the used calibrations and dating_method as attributes(output)$datelife_calibrations and attributes(output)$dating_method.