Use calibrations from each source chronogram to date a tree.

Use calibrations from each source chronogram to date a tree.

use_each_calibration(phy = NULL, phylo_all = NULL, calibrations = NULL, ...)

Arguments

phy
A tree with or without branch lengths to be dated. Only as phylo object for now.

phylo_all
Can be NULL. A datelifeResult list of patristic matrices, or a chronogram as phylo or multiPhylo.

calibrations
Can be NULL. A list of dataframes from get_all_calibrations function.

...
Arguments to pass to get_all_calibrations

Value

A list with a multiPhylo object of chronograms and a list of all calibrations used

Details

You can get the datelifeResult object or the list of chronograms first phylo_all1 <- get_datelife_result(input = my_phy) phylo_all2 <- summarize_datelife_result(phylo_all1) Either will work the same: use_each_calibration(phy = my_phy, phylo_all = phylo_all1) use_each_calibration(phy = my_phy, phylo_all = phylo_all2) You can also get the list of calibrations outside my_calibrations <- get_all_calibrations(input = phylo_all1, each = TRUE) use_each_calibration(phy = my_phy, calibrations = my_calibrations) All this means that you can use your own set of calibrations, not necessarily the ones found only in datelife.