Gets classification paths for a vector of taxa

This uses the taxize package's wrapper of the Global Names Resolver to get taxonomic paths for the vector of taxa you pass in. Sources is a vector of source labels in order (though it works best if everything uses the same taxonomy, so we recommend doing just one source). You can see options by doing taxize::gnr_datasources(). Our default is Catalogue of Life

classification_paths_from_taxonomy(taxa, sources = "Catalogue of Life")

Arguments

taxa
Vector of taxon names

sources
Vector of names of preferred sources; see taxize::gnr_datasources(). Currently supports 100 taxonomic resources, see details.

Value

A list with resolved taxa (a tibble, from taxize::gnr_resolve) and a vector of taxa not resolved

Details

Taxonomies supported by taxize::gnr_datasources()

Catalogue of Life

Wikispecies

ITIS

NCBI

Index Fungorum

GRIN Taxonomy for Plants

Union 4

The Interim Register of Marine and Nonmarine Genera

World Register of Marine Species

Freebase

GBIF Backbone Taxonomy

EOL

Passiflora vernacular names

Inventory of Fish Species in the Wami River Basin

Pheasant Diversity and Conservation in the Mt. Gaoligonshan Region

Finding Species

Birds of Lindi Forests Plantation

Nemertea

Kihansi Gorge Amphibian Species Checklist

Mushroom Observer

TaxonConcept

Amphibia and Reptilia of Yunnan

Common names of Chilean Plants

Invasive Species of Belgium

ZooKeys

COA Wildlife Conservation List

AskNature

China: Yunnan, Southern Gaoligongshan, Rapid Biological Inventories Report No. 04

Native Orchids from Gaoligongshan Mountains, China

Illinois Wildflowers

Coleorrhyncha Species File

/home/dimus/files/dwca/zoological names.zip

Peces de la zona hidrogeográfica de la Amazonia, Colombia (Spreadsheet)

Eastern Mediterranean Syllidae

Gaoligong Shan Medicinal Plants Checklist

birds_of_tanzania

AmphibiaWeb

tanzania_plant_sepecimens

Papahanaumokuakea Marine National Monument

Taiwanese IUCN species list

BioPedia

AnAge

Embioptera Species File

Global Invasive Species Database

Sendoya S., Fernández F. AAT de hormigas (Hymenoptera: Formicidae) del Neotrópico 1.0 2004 (Spreadsheet)

Flora of Gaoligong Mountains

ARKive

True Fruit Flies (Diptera, Tephritidae) of the Afrotropical Region

3i - Typhlocybinae Database

CATE Sphingidae

ZooBank

Diatoms

AntWeb

Endemic species in Taiwan

Dermaptera Species File

Mantodea Species File

Birds of the World: Recommended English Names

New Zealand Animalia

Blattodea Species File

Plecoptera Species File

/home/dimus/files/dwca/clemens.zip

Coreoidea Species File

Freshwater Animal Diversity Assessment - Normalized export

Catalogue of Vascular Plant Species of Central and Northeastern Brazil

Wikipedia in EOL

Database of Vascular Plants of Canada (VASCAN)

Phasmida Species File

OBIS

USDA NRCS PLANTS Database

Catalog of Fishes

Aphid Species File

The National Checklist of Taiwan

Psocodea Species File

FishBase

3i - Typhlocybinae Database

Belgian Species List

EUNIS

CU*STAR

Orthoptera Species File

Bishop Museum

IUCN Red List of Threatened Species

BioLib.cz

Tropicos - Missouri Botanical Garden

nlbif

The International Plant Names Index

Index to Organism Names

uBio NameBank

Arctos

Checklist of Beetles (Coleoptera) of Canada and Alaska. Second Edition.

The Paleobiology Database

The Reptile Database

The Mammal Species of The World

BirdLife International

Checklist da Flora de Portugal (Continental, Açores e Madeira)

FishBase Cache

Silva

Open Tree of Life Reference Taxonomy

iNaturalist

The Interim Register of Marine and Nonmarine Genera

Gymno