Date with Paleobiology Database and paleotree

This will take a topology, look up information about fossils for taxa on the tree, and use paleotree's timePaleoPhy() function to compute branch lengths.

date_with_pbdb(phy, recent = FALSE, assume_recent_if_missing = TRUE)

Arguments

phy

Phylogeny of taxa

recent

If TRUE, forces the minimum age to be zero for any taxon

assume_recent_if_missing

If TRUE, any taxon missing from pbdb is assumed to be recent

Value

A dated tree

Examples

taxa <- c(
  "Archaeopteryx", "Pinus", "Quetzalcoatlus", "Homo sapiens",
  "Tyrannosaurus rex", "Megatheriidae", "Metasequoia", "Aedes", "Panthera"
)
phy <- tree_from_taxonomy(taxa, sources = "The Paleobiology Database")$phy
#> has class
#> reordered
#> collapsed