Core function to input a vector of species, newick string, or phylo object to get a chronogram or dates back.

Core function to input a vector of species, newick string, or phylo object to get a chronogram or dates back.

datelife_search(
  input = c("Rhea americana", "Pterocnemia pennata", "Struthio camelus"),
  summary_format = "phylo_all",
  partial = TRUE,
  use_tnrs = FALSE,
  approximate_match = TRUE,
  update_cache = FALSE,
  cache = getAnywhere("opentree_chronograms"),
  summary_print = c("citations", "taxa"),
  taxon_summary = c("none", "summary", "matrix"),
  get_spp_from_taxon = FALSE,
  verbose = FALSE,
  criterion = "taxa"
)

Arguments

input
A character vector of taxon names, a tree as a 'phylo' object or a newick character string, or a 'datelifeQuery' object from make_datelife_query function.

summary_format
The desired output format for target chronograms (chronograms of target taxa). See details.

partial
If TRUE, use source chronograms even if they only match some of the desired taxa

use_tnrs
If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.

approximate_match
If TRUE, use a slower TNRS to correct misspellings, increasing the chance of matches (including false matches).

update_cache
default to FALSE

cache
The cached set of chronograms and other info from data(opentree_chronograms).

summary_print
A character vector specifying type of summary information to be printed: "citations" for the references of chronograms from cache where target taxa are found, "taxa" for a summary of the number of chronograms where each target taxon is found, or "none" if nothing should be printed. Default to display both c("citations", "taxa").

taxon_summary
A character vector specifying if data on target taxa missing in source chronograms should be added to the output as a "summary" or as a presence/absence "matrix". Default to "none", no information on taxon_summary added to the output.

get_spp_from_taxon
boolean vector, default to FALSE. If TRUE, will get all species names from taxon names given in input. Must have same length as input. If input is a newick string , with some clades it will be converted to phylo object phy, and the order of get_spp_from_taxon will match phy$tip.label.

verbose
Boolean. If TRUE, it gives printed updates to the user.

criterion
Whether to get the grove with the most trees or the most taxa

Details

Available output formats are:

citations: A character vector of references where chronograms with some or all of the target taxa are published (source chronograms).

mrca: A named numeric vector of most recent common ancestor (mrca) ages of target taxa defined in input, obtained from the source chronograms. Names of mrca vector are equal to citations.

newick_all: A named character vector of newick strings corresponding to target chronograms derived from source chronograms. Names of newick_all vector are equal to citations.

newick_sdm: Only if multiple source chronograms are available. A character vector with a single newick string corresponding to a target chronogram obtained with SDM supertree method (Criscuolo et al. 2006).

newick_median: Only if multiple source chronograms are available. A character vector with a single newick string corresponding to a target chronogram from the median of all source chronograms.

phylo_sdm: Only if multiple source chronograms are available. A phylo object with a single target chronogram obtained with SDM supertree method (Criscuolo et al. 2006).

phylo_median: Only if multiple source chronograms are available. A phylo object with a single target chronogram obtained from source chronograms with median method.

phylo_all: A named list of phylo objects corresponding to each target chronogram obtained from available source chronograms. Names of phylo_all list correspond to citations.

phylo_biggest: The chronogram with the most taxa. In the case of a tie, the chronogram with clade age closest to the median age of the equally large trees is returned.

html: A character vector with an html string that can be saved and then opened in any web browser. It contains a 4 column table with data on target taxa: mrca, number of taxa, citations of source chronogram and newick target chronogram.

data_frame A data frame with data on target taxa: mrca, number of taxa, citations of source chronograms and newick string.

For approaches that return a single synthetic tree, it is important that the trees leading to it form a grove (roughly, a sufficiently overlapping set of taxa between trees: see Ané et al. 2005, 10.1007/s00026-009-0017-x). In the rare case of multiple groves, should we take the one with the most trees or the most taxa?

Examples

# obtain median ages from a set of source chronograms in newick format:
ages <- datelife_search(c("Rhea americana", "Pterocnemia pennata", "Struthio camelus",
    "Mus musculus"), summary_format="newick_median")

#> Warning: Datelife Result object is empty.

#> Input taxa were not found across available chronograms.

#> Setting use_tnrs = TRUE might change this, but it is time consuming.

#> Error in `rownames<-`(`*tmp*`, value = paste0("Chronogram", sequence(nrow(taxon_matrix)))): attempt to set 'rownames' on an object with no dimensions

# save the tree in newick format
write(ages, file="some.bird.ages.txt")

#> Error in cat(x, file = file, sep = c(rep.int(sep, ncolumns - 1), "\n"),     append = append): object 'ages' not found

# obtain median ages from a set of source chronograms in phylo format
# will produce same tree as above but in r phylo format:
ages.again <- datelife_search(c("Rhea americana", "Pterocnemia pennata", "Struthio camelus",
    "Mus musculus"), summary_format="phylo_median")

#> Warning: Datelife Result object is empty.

#> Input taxa were not found across available chronograms.

#> Setting use_tnrs = TRUE might change this, but it is time consuming.

#> Error in `rownames<-`(`*tmp*`, value = paste0("Chronogram", sequence(nrow(taxon_matrix)))): attempt to set 'rownames' on an object with no dimensions

plot(ages.again)

#> Error in plot(ages.again): object 'ages.again' not found

library(ape)
ape::axisPhylo()

#> Error in get("last_plot.phylo", envir = .PlotPhyloEnv): object 'last_plot.phylo' not found

mtext("Time (million years ago)", side = 1, line = 2, at = (max(getAnywhere("last_plot.phylo",
    envir = .PlotPhyloEnv)$xx) * 0.5))

#> Error in getAnywhere("last_plot.phylo", envir = .PlotPhyloEnv): unused argument (envir = .PlotPhyloEnv)

write.tree(ages.again, file="some.bird.tree.again.txt") # saves phylo object in newick format

#> Error in write.tree(ages.again, file = "some.bird.tree.again.txt"): object 'ages.again' not found

# obtain mrca ages and target chronograms from all source chronograms
# generate an html  output readable in any web browser:
ages.html <- datelife_search(c("Rhea americana", "Pterocnemia pennata", "Struthio camelus",
    "Mus musculus"), summary_format="html")

#> Warning: Datelife Result object is empty.

#> Input taxa were not found across available chronograms.

#> Setting use_tnrs = TRUE might change this, but it is time consuming.

#> Error in `rownames<-`(`*tmp*`, value = paste0("Chronogram", sequence(nrow(taxon_matrix)))): attempt to set 'rownames' on an object with no dimensions

write(ages.html, file="some.bird.trees.html")

#> Error in cat(x, file = file, sep = c(rep.int(sep, ncolumns - 1), "\n"),     append = append): object 'ages.html' not found

system("open some.bird.trees.html")