Get ott ids from a character vector containing species names and ott ids.

Get ott ids from a character vector containing species names and ott ids.

extract_ott_ids(x, na.rm = TRUE)

# S3 method for default
extract_ott_ids(x, na.rm = TRUE)

Arguments

x

A character vector of taxon names, or a phylo object with tip names containing ott ids.

na.rm

A logical value indicating whether NA values should be stripped in the output.

Value

An object of class numeric containing ott ids only. NULL

Examples

canis <- rotl::tnrs_match_names("canis")
canis_taxonomy <- rotl::taxonomy_subtree(canis$ott_id)
my_ott_ids <- extract_ott_ids(x = canis_taxonomy$tip_label)
#> After extracting ott ids, there are some non numeric elements:
#> 	 Canissp.KEB-2016ott5925604
#> 	 'Canissp.Russia/33
#> 	 500ott5338950'
#> 	 'Canissp.Belgium/36
#> 	 000ott5338951'
#> 
#> NAs removed.
# get the problematic elements from input
canis_taxonomy$tip_label[attr(my_ott_ids, "na.action")]
#> [1] "Canissp.KEB-2016ott5925604" "'Canissp.Russia/33"        
#> [3] "500ott5338950'"             "'Canissp.Belgium/36"       
#> [5] "000ott5338951'"