Gets the ott id and name of all lineages from one or more input taxa

Gets the ott id and name of all lineages from one or more input taxa

get_ott_lineage(input = NULL, ott_ids = NULL)

Arguments

input
Optional. A character vector of names or a datelifeQuery object

ott_ids
If not NULL, it takes this argument and ignores input. A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info or rolt::tnrs_match_names or datelife::tnrs_match

Value

a list of named numeric vectors of ott ids from input and all the clades it belongs to.

Examples

taxa <- c("Homo", "Bacillus anthracis", "Apis", "Salvia")
lin <- get_ott_lineage(taxa)

#> Running 'make_datelife_query'...

#>   |                                                                              |                                                                      |   0%  |                                                                              |==================                                                    |  25%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================================                  |  75%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |==================                                                    |  25%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================================                  |  75%  |                                                                              |======================================================================| 100%

lin

#> $Homo
#>                                               ott_ids   ott_ranks    
#> Homininae                                     "312031"  "subfamily"  
#> Hominidae                                     "770311"  "family"     
#> Hominoidea                                    "386191"  "superfamily"
#> Catarrhini                                    "842867"  "parvorder"  
#> Simiiformes                                   "386195"  "infraorder" 
#> Haplorrhini                                   "702152"  "suborder"   
#> Primates                                      "913935"  "order"      
#> Euarchontoglires                              "392222"  "superorder" 
#> Boreoeutheria                                 "5334778" "no rank"    
#> Eutheria (in Deuterostomia)                   "683263"  "no rank"    
#> Theria (subclass in Deuterostomia)            "229558"  "subclass"   
#> Mammalia                                      "244265"  "class"      
#> Amniota                                       "229560"  "no rank"    
#> Tetrapoda                                     "229562"  "superclass" 
#> Dipnotetrapodomorpha                          "4940726" "no rank"    
#> Sarcopterygii                                 "458402"  "superclass" 
#> Euteleostomi                                  "114654"  "no rank"    
#> Teleostomi                                    "114656"  "no rank"    
#> Gnathostomata (superclass in phylum Chordata) "278114"  "superclass" 
#> Vertebrata (subphylum in Deuterostomia)       "801601"  "subphylum"  
#> Craniata (subphylum in Deuterostomia)         "947318"  "subphylum"  
#> Chordata                                      "125642"  "phylum"     
#> Deuterostomia                                 "147604"  "no rank"    
#> Bilateria                                     "117569"  "no rank"    
#> Eumetazoa                                     "641038"  "no rank"    
#> Metazoa                                       "691846"  "kingdom"    
#> Holozoa                                       "5246131" "no rank"    
#> Opisthokonta                                  "332573"  "no rank"    
#> Eukaryota                                     "304358"  "domain"     
#> cellular organisms                            "93302"   "no rank"    
#> life                                          "805080"  "no rank"    
#> 
#> $`Bacillus anthracis`
#>                                     ott_ids   ott_ranks
#> Bacillus (genus in domain Bacteria) "1068896" "genus"  
#> Bacillaceae                         "589414"  "family" 
#> Bacillales                          "1068893" "order"  
#> Bacilli                             "634420"  "class"  
#> Firmicutes                          "822769"  "phylum" 
#> Bacteria                            "844192"  "domain" 
#> cellular organisms                  "93302"   "no rank"
#> life                                "805080"  "no rank"
#> 
#> $Apis
#>                                      ott_ids   ott_ranks    
#> Apini                                "806703"  "tribe"      
#> Apinae                               "849856"  "subfamily"  
#> Apidae                               "52848"   "family"     
#> Apoidea                              "419526"  "superfamily"
#> Aculeata (infraorder in Protostomia) "206138"  "infraorder" 
#> Apocrita                             "678075"  "suborder"   
#> Hymenoptera                          "753726"  "order"      
#> Holometabola                         "1082885" "no rank"    
#> Neoptera                             "815350"  "infraclass" 
#> Pterygota (subclass in Opisthokonta) "1048707" "subclass"   
#> Dicondylia                           "983656"  "no rank"    
#> Insecta                              "1062253" "class"      
#> Hexapoda                             "568991"  "subphylum"  
#> Pancrustacea                         "985906"  "no rank"    
#> Mandibulata                          "985907"  "no rank"    
#> Arthropoda                           "632179"  "phylum"     
#> Panarthropoda                        "816442"  "no rank"    
#> Ecdysozoa                            "611099"  "no rank"    
#> Protostomia                          "189832"  "no rank"    
#> Bilateria                            "117569"  "no rank"    
#> Eumetazoa                            "641038"  "no rank"    
#> Metazoa                              "691846"  "kingdom"    
#> Holozoa                              "5246131" "no rank"    
#> Opisthokonta                         "332573"  "no rank"    
#> Eukaryota                            "304358"  "domain"     
#> cellular organisms                   "93302"   "no rank"    
#> life                                 "805080"  "no rank"    
#> 
#> $Salvia
#>                      ott_ids   ott_ranks  
#> Mentheae             "498490"  "tribe"    
#> Nepetoideae          "394658"  "subfamily"
#> Lamiaceae            "544714"  "family"   
#> higher_core_Lamiales "5264061" "no rank"  
#> Core_Lamiales        "5263556" "no rank"  
#> Lamiales             "23736"   "order"    
#> lamiids              "596112"  "no rank"  
#> asterids             "1008294" "no rank"  
#> Pentapetalae         "5316182" "no rank"  
#> Gunneridae           "853757"  "no rank"  
#> eudicotyledons       "431495"  "no rank"  
#> Mesangiospermae      "5298374" "no rank"  
#> Magnoliopsida        "99252"   "class"    
#> Spermatophyta        "10218"   "no rank"  
#> Euphyllophyta        "1007992" "no rank"  
#> Tracheophyta         "10210"   "phylum"   
#> Embryophyta          "56610"   "no rank"  
#> Streptophyta         "916750"  "phylum"   
#> Chloroplastida       "361838"  "kingdom"  
#> Archaeplastida       "5268475" "kingdom"  
#> Eukaryota            "304358"  "domain"   
#> cellular organisms   "93302"   "no rank"  
#> life                 "805080"  "no rank"  
#> 

 # look up an unknown ott id:
 get_ott_lineage(ott_id = 454749)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> $`Discelium nudum`
#>                    ott_ids   ott_ranks 
#> Discelium          "508293"  "genus"   
#> Disceliaceae       "538343"  "family"  
#> Funariales         "821347"  "order"   
#> Funariidae         "620630"  "subclass"
#> Bryopsida          "821346"  "class"   
#> Bryophytina        "471195"  "no rank" 
#> Bryophyta          "246594"  "phylum"  
#> Embryophyta        "56610"   "no rank" 
#> Streptophyta       "916750"  "phylum"  
#> Chloroplastida     "361838"  "kingdom" 
#> Archaeplastida     "5268475" "kingdom" 
#> Eukaryota          "304358"  "domain"  
#> cellular organisms "93302"   "no rank" 
#> life               "805080"  "no rank" 
#>