Use genetic data from the Barcode of Life Database (BOLD) to reconstruct branch lengths on a tree.

make_bold_otol_tree takes taxon names from a tree topology or a vector of names to search for genetic markers in the Barcode of Life Database (BOLD), create an alignment, and reconstruct branch lengths on a tree topology with Maximum Likelihood.

make_bold_otol_tree(
  input = c("Rhea americana", "Struthio camelus", "Gallus gallus"),
  marker = "COI",
  otol_version = "v3",
  chronogram = TRUE,
  doML = FALSE,
  aligner = "muscle",
  ...
)

Arguments

input

One of the following:

A character vector

With taxon names as a single comma separated strting or concatenated with c().

A phylogenetic tree with taxon names as tip labels

As a phylo or multiPhylo object, OR as a newick character string.

A datelifeQuery object

An output from make_datelife_query().

marker

A character vector indicating the gene from BOLD system to be used for branch length estimation.

otol_version

Version of Open Tree of Life to use

chronogram

Default to TRUE, branch lengths returned are estimated with ape::chronoMPL(). If FALSE, branch lengths returned are estimated with phangorn::acctran() and represent relative substitution rates.

doML

Default to FALSE. If TRUE, it does ML branch length optimization with phangorn::optim.pml(). Only relevant if chronogram = TRUE.

aligner

A character vector indicating whether to use MAFFT or MUSCLE to align BOLD sequences. It is not case sensitive. Default to MUSCLE, supported using the msa package from Bioconductor. Needs to be installed using BiocManager::install().

...

Arguments passed on to get_otol_synthetic_tree

resolve

boolean; default to TRUE. Whether to resolve the tree at random or not.

ott_ids

If not NULL, it takes this argument and ignores input. A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info() or rotl::tnrs_match_names() or tnrs_match()

Value

A phylogeny with branch lengths poportional to relative substitution rate.

Details

If input is a phylo object or a newick string, it is used as backbone topology. If input is a character vector of taxon names, an induced synthetic OpenTree subtree is used as backbone. If there are not enough sequences to return a tree with branch lengths, it returns either the original input topology or the synthetic OpenTree subtree obtained for the input taxon names.