Summarize a filtered results list from get_datelife_result function in various ways

Summarize a filtered results list from get_datelife_result function in various ways

summarize_datelife_result(datelife_query = NULL,
  datelife_result = NULL, summary_format = "phylo_all",
  partial = TRUE, update_cache = FALSE,
  cache = get("opentree_chronograms"), summary_print = c("citations",
  "taxa"), taxon_summary = c("none", "summary", "matrix"),
  verbose = FALSE, criterion = c("trees", "taxa"))

Arguments

datelife_query
A datelifeQuery object, output of make_datelife_query function

datelife_result
An output of get_datelife_result function: A list of patristic matrices with names corresponding to original study citations.

summary_format
The desired output format for target chronograms (chronograms of target taxa). See details.

partial
If TRUE, use source chronograms even if they only match some of the desired taxa

update_cache
default to FALSE

cache
The cached set of chronograms and other info from data(opentree_chronograms).

summary_print
A character vector specifying type of summary information to be printed: "citations" for the references of chronograms from cache where target taxa are found, "taxa" for a summary of the number of chronograms where each target taxon is found, or "none" if nothing should be printed. Default to display both c("citations", "taxa").

taxon_summary
A character vector specifying if data on target taxa missing in source chronograms should be added to the output as a "summary" or as a presence/absence "matrix". Default to "none", no information on taxon_summary added to the output.

verbose
Boolean. If TRUE, it gives printed updates to the user.

criterion
Whether to get the grove with the most trees or the most taxa

Details

Available output formats are:

citations: A character vector of references where chronograms with some or all of the target taxa are published (source chronograms).

mrca: A named numeric vector of most recent common ancestor (mrca) ages of target taxa defined in input, obtained from the source chronograms. Names of mrca vector are equal to citations.

newick_all: A named character vector of newick strings corresponding to target chronograms derived from source chronograms. Names of newick_all vector are equal to citations.

newick_sdm: Only if multiple source chronograms are available. A character vector with a single newick string corresponding to a target chronogram obtained with SDM supertree method (Criscuolo et al. 2006).

newick_median: Only if multiple source chronograms are available. A character vector with a single newick string corresponding to a target chronogram from the median of all source chronograms.

phylo_sdm: Only if multiple source chronograms are available. A phylo object with a single target chronogram obtained with SDM supertree method (Criscuolo et al. 2006).

phylo_median: Only if multiple source chronograms are available. A phylo object with a single target chronogram obtained from source chronograms with median method.

phylo_all: A named list of phylo objects corresponding to each target chronogram obtained from available source chronograms. Names of phylo_all list correspond to citations.

phylo_biggest: The chronogram with the most taxa. In the case of a tie, the chronogram with clade age closest to the median age of the equally large trees is returned.

html: A character vector with an html string that can be saved and then opened in any web browser. It contains a 4 column table with data on target taxa: mrca, number of taxa, citations of source chronogram and newick target chronogram.

data_frame A data frame with data on target taxa: mrca, number of taxa, citations of source chronograms and newick string.

For approaches that return a single synthetic tree, it is important that the trees leading to it form a grove (roughly, a sufficiently overlapping set of taxa between trees: see Ané et al. 2005, 10.1007/s00026-009-0017-x). In the rare case of multiple groves, should we take the one with the most trees or the most taxa?