Go from a summary matrix to an ultrametric phylo object.

Go from a summary matrix to an ultrametric phylo object.

summary_matrix_to_phylo(
  summ_matrix,
  datelife_query = NULL,
  total_distance = TRUE,
  use = "mean",
  target_tree = NULL,
  ...
)

Arguments

summ_matrix
A summary patristic distance matrix from sdm or median. See details.

datelife_query
A datelifeQuery object, output of make_datelife_query function

total_distance
Boolean. If TRUE it will divide the matrix in half, if FALSE it will take it as is.

use
A character vector indicating what type of age to use for summary. One of the following

mean

It will use the mean of the node age distributions.

min

It will use the minimum age from the node age distributions.

max

Choose this if you wanna be conservative; it will use the maximum age from the node age distributions.

target_tree
A phylo object. Use this in case you want a particular backbone for the output tree.

...
Arguments passed on to get_otol_synthetic_tree

otol_version

Version of OToL to use

resolve

boolean default to TRUE. Whether to resolve the tree at random or not.

input

Optional. A character vector of names or a datelifeQuery object

ott_ids

If not NULL, it takes this argument and ignores input. A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info or rolt::tnrs_match_names or datelife::tnrs_match

Value

An ultrametric phylo object.

Details

It can take a regular patristic distance matrix, but there are simpler methods for that implemented in patristic_matrix_to_phylo.